Guard thread local storage by mutex.
Currently, ParserContext is implemented via a thread local storage, but
it has not been thread safe.

The bug is not really effective, as the parser runs in a single thread.
We can even use global storage for parser contexts. However, my feeling
is that it is a bit creepy to leave global storage that can trivially be
made thread-safe.