warn when a jpeg image with CMYK color mode is used
Now SATysFi can load a jpeg image with CMYK color mode, but it cannot output a proper image.
As a temporary solution, a warning that notifies the fact that one is using an image with CMYK color model is added.

When a jpeg image with CMYK color mode is used SATySFi outputs an image with wrong colors.
But this behaviour is better than before; before the image was losing its shape.
Probably, the problem is caused because SATySFi doesn't embed a decode array into the pdf.

I had to modify `pass_type_check` in `logging.ml` to avoid circular dependency.